+++ src/EasyWinGetPlus.Launcher.cs
@@ -9,12 +9,12 @@ using System.Windows.Forms;
 [assembly: AssemblyCompany("廖阿輝")]
 [assembly: AssemblyProduct("Easy WinGet Plus")]
 [assembly: AssemblyCopyright("Copyright © 2026 廖阿輝")]
-[assembly: AssemblyVersion("0.1.0.0")]
-[assembly: AssemblyFileVersion("0.1.0.0")]
+[assembly: AssemblyVersion("0.1.2.0")]
+[assembly: AssemblyFileVersion("0.1.2.0")]
 
 internal static class Program
 {
-    private const string Version = "0.1.0";
+    private const string Version = "0.1.2";
     private const string ResourceName = "EasyWinGetPlus.ps1";
 
     [STAThread]